Move getMenus to functions
